[runtime] Allow to use a dynamic event dispatcher with @dynamicEvents meta on a class
This is meant to be used when we prefer limiting the number of generated methods in class, at the cost of more dynamic calls which may be more difficult to track. This is a required feature to make ceramic work with targets like lua eventually. (the number of fields per class in lua output is limited)