[tools] Allow to replace haxe std files with custom ones provided by ceramic in /haxe/std
This allows to patch haxe’s standard library in cases where it is not possible to do otherwise to make ceramic work properly. Ultimately, any change made to standard library should become a pull request to official standard library, but this allows to apply and use changes immediatly in a reproducible way.