[backend] Stop using luxe for saving/reading strings on web. Directly use localStorage because it is more efficient.
This was needed because luxe is performing unnecessary additional encoding of the strings we pass, and does it for every strings of the same slot everytime. When managing a lot of keys, this was freezing the app at each change. Using localStorage directly provides the same set of features we need, minus the overhead.