[runtime] Event owner argument is now required. Previous behavior possible with ceramic_optional_owner flag but strongly discouraged
We are enforcing the rule: never forget to free/unbind any object. Specifying an event owner is helping that. Making it optional could lead to developer forgetting to specify an owner when he should. It is still possible to specify « no owner » by providing null as owner, we are just making things more explicit, and choosing the safest option as the default