[runtime] Component is now an interface. Entity can be instanciated as is.
This is another breaking change, but very useful as it will provide much more flexibility with what can be used as Component. As long as an object implements Component and inherits Entity, it can now be used as a component. That opens the door to optionally use first-class objects as component bound to other entities.